Full option !!! 90% without 3 items

+++ Catch me if you can/core/src/com/mygdx/game/tools/Clipping.java
@@ -3,8 +3,8 @@ package com.mygdx.game.tools;
 import com.badlogic.gdx.Gdx;
 
 public class Clipping {
-	public static int check_x(int player_x, int size_x) {
-		int x = player_x;
+	public static float check_x(float player_x, int size_x) {
+		float x = player_x;
 		if (x > 1280 - size_x)
 		{
 			x = 1280 - size_x;
@@ -15,8 +15,8 @@ public class Clipping {
 		}
 		return x;
 	}
-	public static int check_y(int player_y, int size_y) {
-		int y = player_y;
+	public static float check_y(float player_y, int size_y) {
+		float y = player_y;
 		if (y > 680 - size_y)
 		{
 			y = 680 - size_y;
